complete snode

+++ program/compile_config.h
@@ -6,8 +6,9 @@
 namespace redwood::lang {
 
 struct CompileConfig {
-  /*
+  
   Arch arch;
+  /*
   bool debug;
   bool cfg_optimization;
   bool check_out_of_bound;
@@ -26,7 +27,9 @@ struct CompileConfig {
   bool simplify_after_lower_access;
   bool move_loop_invariant_outside_if;
   bool cache_loop_invariant_global_vars{true};
+  */
   bool demote_dense_struct_fors;
+  /*
   bool advanced_optimization;
   bool constant_folding;
   bool use_llvm;
@@ -46,20 +49,24 @@ struct CompileConfig {
   DataType default_ip;
   DataType default_up;
   std::string extra_flags;
+  */
   int default_cpu_block_dim;
   bool cpu_block_dim_adaptive;
   int default_gpu_block_dim;
+  /*
   int gpu_max_reg;
   int ad_stack_size{0};  // 0 = adaptive
   // The default size when the redwood compiler is unable to automatically
   // determine the autodiff stack size.
   int default_ad_stack_size{32};
-
+  */
   int saturating_grid_dim;
+  
   int max_block_dim;
   int cpu_max_num_threads;
+  
   int random_seed;
-
+  /*
   // LLVM backend options:
   bool print_struct_llvm_ir;
   bool print_kernel_llvm_ir;
@@ -68,9 +75,10 @@ struct CompileConfig {
   bool print_kernel_amdgcn;
 
   // CUDA/AMDGPU backend options:
-  float64 device_memory_GB;
-  float64 device_memory_fraction;
-
+  */
+  float device_memory_GB;
+  float device_memory_fraction;
+  /*
   // Opengl backend options:
   bool allow_nv_shader_extension{true};
 
@@ -96,12 +104,12 @@ struct CompileConfig {
   int offline_cache_max_size_of_files{100 * 1024 *
                                       1024};   // bytes, default: 100MB
   double offline_cache_cleaning_factor{0.25};  // [0.f, 1.f]
-
+  */
   int num_compile_threads{4};
   std::string vk_api_version;
 
   size_t cuda_stack_limit{0};
-  */
+  
   CompileConfig();
 
   //void fit();
